7_Tamara_Agent/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.0.5
python-snap7>=1.3
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.7.0
chromadb>=0.4.22
pytest>=8.0.0
pydantic>=2.7.0
